Move the speech locale cache into an actor
Serializes the probe regardless of the caller's isolation, so only one
runs per process, and looks the selected voice up by identifier instead
of enumerating the whole catalog.
